+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/app
 
+## 2.0.0-beta.11
+
+### Patch Changes
+
+- s
+
 ## 2.0.0-beta.10
 
 ### Patch Changes