+++ apps/mini-c3-electron/CHANGELOG.md
@@ -1,5 +1,11 @@
 # mini-c3-electron
 
+## 0.0.1-beta.2
+
+### Patch Changes
+
+- s
+
 ## 0.0.1-beta.1
 
 ### Patch Changes