+++ apps/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/ui
 
+## 2.0.0-beta.14
+
+### Patch Changes
+
+- s
+- Updated dependencies
+  - @pipelab/constants@1.0.0-beta.12
+  - @pipelab/shared@1.0.0-beta.10
+
 ## 2.0.0-beta.13
 
 ### Patch Changes