+++ assets/asset-electron/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-electron
 
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- s
+
 ## 1.0.0-beta.9
 
 ### Patch Changes