+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/constants
 
+## 1.0.0-beta.12
+
+### Patch Changes
+
+- s
+
 ## 1.0.0-beta.11
 
 ### Patch Changes